Downgrade mingw-w64-osslsigncode to an OpenSSL v1.0.2 version
MSYS2 started to upgrade OpenSSL to v1.1.1. This breaks tons of
dependencees, as MSYS2 did not even plan for side-by-side installations,
and the version name of the OpenSSL library is embedded in the .dll
(meaning: every major update will break dependencies).

We will have to do this carefully in Git for Windows, quite a bit more
carefully than MSYS2, and for the time being, will roll back the
osslsigncode update to unbreak things.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>