Work around the http.sslCAInfo setting now being shared between MSYS/MINGW
In Git for Windows, we actually no longer need that setting, but we use
it still because Git LFS needs it to determine whether to override the
certificates of the Windows Credential Store or not.

Since this setting pretends to be an absolute path, but actually is
relative to Git's runtime prefix, its value looks like this:
`/ssl/certs/ca-bundle.crt`. In the MSYS context, however, this is wrong:
it should be `/usr/ssl/certs/ca-bundle.crt`.

Let's work around this by providing a symbolic link.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>