Add a GitHub Action to upload some artifacts to Azure Blobs
In Azure Pipelines, we publish a couple Build Artifacts for use by other
Azure Pipelines:

- full-sdk (x86_64 and i686)
- build-installers (x86_64 and i686)
- makepkg-git (x86_64 only)
- minimal (x86_64 only)

As we want to switch to GitHub Actions, we need a different way to
access those artifacts. The solution we settled on is to download
`.tar.xz` files from Azure Blobs and simultaneously extract them, in a
Bash step (which uses Git for Windows' Git Bash, which comes with the
required `curl`, `tar` and `xz`) essentially by calling

        curl <url> | tar xJf -

With this commit, we generate and upload the i686 ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>