Unbreak GPG's `dirmngr`
A recent upgrade to libhogweed and to libnettle broke libgnutls, which
is a dependency of `dirmngr.exe`. As a consequence, GPG does not work.

To work around this (especially to enable us to build a new libgnutls),
let's cherry-pick the now-missing `.dll` files.

Once we have a new and working libgnutls, we will remove them again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>