libtool: work around libpthread not being a DLL
When linking `-lpthread` to a shared library, `libtool` wants to make
sure that we're not linking to a `libpthread.a` when we should link to a
`libpthread.dll.a` instead.

However, the pthread functionality is part of the MSYS2 runtime and not
in a separate `.dll` file, hence it is totally legitimate that we're
linking to a `.a` instead of a `.dll.a` file, and `libtool` is wrong to
complain thusly:

	*** Warning: linker path does not have real file for library -lpthread.
	*** I have the capability to make that library automatically link in when
	*** you link to this library.  But I can only do this if you have a
	*** shared version of the library, which you do not appear to have
	*** because I did check the linker path looking for a file starting
	*** with libpthread and none of the candidates passed a file format test
	*** using a file magic. Last file checked: /usr/lib/libpthread.a
	*** The inter-library dependencies that have been dropped here will be
	*** automatically added whenever a program is linked with this library
	*** or is declared to -dlopen it.

	*** Since this library must not contain undefined symbols,
	*** because either the platform does not support them or
	*** it was explicitly requested with -no-undefined,
	*** libtool will only create a static version of it.

Let's work around that.

Note: this is a pretty dirty hack, and only in the i686 version of Git
for Windows' SDK (because `libtool`'s code targets i686 specifically,
for some reason this developer does not understand). If the need should
arise, we will release a new version of the `libtool` package with this
workaround made more permanent; However, given that it affects i686
only, and given that MSYS2 deprecated i686 and does not release any new
MSYS packages for that architecture (see
https://www.msys2.org/news/#2020-05-17-32-bit-msys2-no-longer-actively-supported
for details) this here patch might be Good Enough.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>