Add a GitHub workflow to synchronize the repository
While git-sdk-32 is a source of truth when it comes to Git for Windows'
SDK, it has to be synchronized with the Pacman repositories of the MSYS2
and the Git for Windows projects.

This was previously done in an Azure Pipeline at
https://dev.azure.com/Git-for-Windows/git-sdk-64/_build/index?definitionId=9,
and is now done in this shiny new GitHub workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>