sync: allow other workflows to be triggered by push
When a GitHub workflow pushes to the repository using the regular token
assigned to the workflow run, no `push` event is sent (to avoid trigger
cascades where workflow runs trigger each other ad infinitum).

But we do want the `ci-artifacts` workflow to be triggered. Therefore,
we use a PAT stored in the `sync` environment to perform that push.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

g Your branch is ahead of 'github/main' by 1 commit.