sync: avoid problems when updating the MSYS2 runtime or Bash
In the original Azure Pipeline "SyncSDKs", we went out of our way to
execute the script `please.sh` using _another_ Bash than the one in Git
for Windows' SDK. The reason is a Catch-22: if the Bash needs to be
updated that runs the script that updates the Bash, it won't work
because on Windows you cannot overwrite files that are still being used.

This careful Bash dance unfortunately got lost when transitioning to a
GitHub workflow, so let's reinstate it here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>