sync: do not trust the mtime of the Pacman database files
Pacman will only update its locally-cached databases by default when
their local timestamps indicate that they are not up to date with the
remote versions.

However, in the GitHub workflow, we check out those files afresh,
meaning: the timestamp will be misleading.

To counter that, we ask Pacman to ignore the local timestamp and force
an update of the locally-cached database files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>