gnutls: work around incorrect DLL prefix
With the `cyg` prefix of the gnutls DLL, GPG's `dirmngr` does not even
start, so we cannot even rebuild gnutls...

Let's work around that by this hack: copy the `cyggnutls-30.dll` to
`msys-gnutls-30.dll`. Let's also add it to the file list so it gets
removed when upgrading to the gnutls package we're about to build.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>