ci: continuously build Git for Windows' artifacts
This hopefully will prevent last-minute surprises on "Release Day"
(which thankfully haven't been all that common as of late).

This is a heavily edited copy of
https://github.com/git-for-windows/git/blob/v2.37.2.windows.2/.github/workflows/git-artifacts.yml

It would have been nicer to be able to reuse parts of that workflow, of
course, but the fact that we must not use `setup-git-for-windows-sdk` in
this workflow (for obvious reasons) makes this not worth the effort.

To simplify the workflow definition, all GPG and code-signing parts have
been dropped: we are unlikely to ever want to advertise these artifacts
as anything remotely official.

In contrast to the original, we do not go for `build-installers` flavors
of the Git for Windows SDK because the 32-bit version of that flavor is
not actually set up to build `mingw-w64-git` (that's done in the 64-bit
flavor in Git for Windows' release process). So we simply go for a
shallow clone of `git-sdk-32` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>