makepkg: fix compatibility with bash-5.2 globskipdots
Bash 5.2 has a new globskipdots option, which is enabled by default. The
check_dotfiles lint fails with globskipdots due to the assumption that
at least the "." and ".." paths will match. Disabling globskipdots would
be the usual solution, but that fails on bash<5.2.  Instead, enable
nullglob for this check.

This is a backport of this commit (which made it into Pacman v6.0.2):
https://gitlab.archlinux.org/pacman/pacman/-/commit/5c10680f6f5f9ed1eb456f57cba3f88ecb69066c

Signed-off-by: Allan McRae <allan@archlinux.org>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>