nghttp2: (temporarily) reinstate unversioned DLL
The old DLL encoded the version number in the file name and therefore
dependencees were broken (e.g. cURL).

In https://github.com/msys2/MINGW-packages/pull/13810, due to "helpful"
feedback of users who insist on using an unsupported scenario, the name
was changed back, which made this problem infinitely worse as the
`pacman` update failed due to a file that would be overwritten _and_
breaking cURL again.

Let's do the reverse dance, then, and hope that it won't use up as much
time to get back to a working system as it did a couple days ago.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>