sync: decouple from the `build-extra` repository
It is a bit heavy-handed to clone the entire `build-extra` repository
just to perform two tasks, i.e. updating via Pacman and then committing
the result.

In the sibling repository `git-sdk-arm64` we already had to decouple the
`sync` workflow from `please.sh sync` (because the latter's assumptions
could not be fulfilled), and we can adapt that work slightly and
integrate it into this here repository.

This commit copy/edits files from 17723bad53b5 (ci: avoid cloning
`build-extra`, 2023-01-25) in
https://github.com/git-for-windows/git-sdk-arm64, partially addressing
https://github.com/git-for-windows/git/issues/4234

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>