git-artifacts: align with git-sdk-64's version of the workflow
In the i686 version, we used to do things differently because we did not
have a working sparse-checkout definition for the `makepkg-git` SDK
artifact, and therefore we simply used the full SDK.

However, the plan is to actually use the `build-installers` SDK artifact
(which extends the `makepkg-git` SDK artifact) to build the official Git
for Windows releases not only for x86_64, but also for i686 (and for
that matter, also for aarch64).

So let's copy/edit as much as possible from git-sdk-64's
`git-artifacts.yml`.

Note: there are of course still some (intended) deviations:

- we leave the `MSYSTEM` environment variable alone

- we leave the `ARCH_*` variables alone

- when testing the installed Git for Windows, we still expect the i686
  variant to be in `C:\Program Files (x86)`, not in `C:\Program Files`

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>