Hold back broken `libgpg-error` update
The most recent i686 variant of the `libgpg-error` package is broken: by
mistake, it built a `cyggpg-error-0.dll` file instead of an
`msys-gpg-error-0.dll` one.

This breaks everything that uses `gpg`, including any package updates
(because the package database is first verified using `gpg`).

In preparation for building a non-broken update of that package, prevent
updates to the still-broken version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>