Temporarily reinstate obsolete OpenSSL DLLs
Gah. I thought that we used MINGW cURL and was _very_ careful to go
through the motions so that `curl.exe` does not link to OpenSSL by
default.

But of course `makepkg` uses _MSYS_ cURL. And that's now broken and I
cannot even build a new cURL without reinstating the old OpenSSL DLLs.

So let's do that. *sigh*.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>