PR build (run the installer): do install the 32-bit version
Ever since c800a3f60 (installer: loudly warn when installing a 32-bit
Git for Windows, 2023-02-21), we complained loudly when a user tried to
install a 32-bit Git for Windows on a 64-bit Windows.

As somewhat intended side effect, it became impossible to install a
32-bit Git for Windows on a 64-bit Windows.

One of the unintended consequences, though, was that the automated
workflow runs that verify the installer could no longer run the 32-bit
installer successfully.

Unfortunately, due to a bug in the workflow definition (which will be
fixed after this here commit), this problem never caused the workflow
runs to fail, and therefore both bugs went undetected for months.

This commit fixes the bug where the 32-bit installer would not complete
successfully in the workflow runs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>