Downgrade mingw-w64-git to allow big fetches again
The `git-artifacts` workflow failed in a step that tried to fetch
git-for-windows/git's `main` branch.

The reason is this mimalloc bug:
https://github.com/microsoft/mimalloc/issues/1025

Unfortunately, I cannot deploy a fixed version with the current
git-sdk-32 revision because it would not be able to fetch the fixed
version...

So let's downgrade `mingw-w64-i686-git` in preparation for upgrading it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>